+++ src/Config/Immutable.php
@@ -5,37 +5,7 @@
 
 namespace Mvc5\Config;
 
-use Exception;
-
-trait Immutable
+interface Immutable
+    extends Base
 {
-    /**
-     *
-     */
-    use Config {
-        remove as private;
-        set as private;
-    }
-
-    /**
-     * @param mixed $name
-     * @param mixed $value
-     * @return mixed $value
-     * @throws Exception
-     */
-    function offsetSet($name, $value)
-    {
-        throw new Exception('Invalid operation: object cannot be modified');
-    }
-
-    /**
-     * @param mixed $name
-     * @param mixed $value
-     * @return mixed
-     * @throws Exception
-     */
-    function __set($name, $value)
-    {
-        throw new Exception('Invalid operation: object cannot be modified');
-    }
 }
