+++ src/Immutable.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ *
+ */
+
+namespace Mvc5;
+
+class Immutable
+    implements Config\Immutable
+{
+    /**
+     *
+     */
+    use Config\ReadOnly;
+}
