--- src/Object.php
@@ -1,15 +1,0 @@
-<?php
-/**
- *
- */
-
-namespace Mvc5;
-
-class Object
-    implements Config\Object
-{
-    /**
-     *
-     */
-    use Config\Immutable;
-}
